requirements.txt
streamlit
pypdf
sentence-transformers
numpy
python-dotenv
google-generativeai
pytest